[arch] minor tweak for vsf_systimer_tick_to_us/vsf_systimer_tick_to_ms in tick mode

+++ source/hal/arch/vsf_arch_abstraction.c
@@ -937,7 +937,7 @@ WEAK(vsf_systimer_tick_to_us)
 vsf_systimer_tick_t vsf_systimer_tick_to_us(vsf_systimer_tick_t tick)
 {
     if (tick > 0) {
-        uint64_t tmp = tick * __systimer.cycle_per_tick * 1000000;
+        uint64_t tmp = (uint64_t)tick * __systimer.cycle_per_tick * 1000000;
         tmp /= vsf_arch_req___systimer_freq___from_usr();
         tick = vsf_max(1, tmp);
     }
@@ -948,7 +948,7 @@ WEAK(vsf_systimer_tick_to_ms)
 vsf_systimer_tick_t vsf_systimer_tick_to_ms(vsf_systimer_tick_t tick)
 {
     if (tick > 0) {
-        uint64_t tmp = tick * __systimer.cycle_per_tick * 1000;
+        uint64_t tmp = (uint64_t)tick * __systimer.cycle_per_tick * 1000;
         tmp /= vsf_arch_req___systimer_freq___from_usr();
         tick = vsf_max(1, tmp);
     }
